Also keep any Addressable Resource manager
As this is a Unity Package I think it should be added here (although in the long run there should be a way to specify additional files to keep depending on packages/own services). I pulled my hair out on this one. My tests where basically just hanging in the void because async operations where not finishing anymore due to this GO suppression.